Update aurora.sh
checks for /usr/share/aurora/.UNRESIZED on startup and executes resize.sh if it's not present
the resize.sh code was already written, someone just either accidentally removed or forgot to add it to startup if i had to guess. it works now though! i think..
this is my first proper commit, sorry if it sucks TvT